refactor: controller 하나로 통합
